+++ server/component/style/qualtricsSurvey/QualtricsSurveyModel.php
@@ -145,12 +145,9 @@ class QualtricsSurveyModel extends StyleModel
         if ($survey_info) {
             $survey_link =  'https://eu.qualtrics.com/jfe/form/' . $survey_info['qualtrics_survey_id'];
             if (isset($survey_info['participant_variable']) && $survey_info['participant_variable'] != '') {
-                $user_code = $this->db->query_db_first('SELECT code
-                                        FROM validation_codes vc
-                                        INNER JOIN users u ON (u.id = vc.id_users)
-                                        WHERE u.id = :id', array(':id' => $_SESSION['id_user']));
-                if (isset($user_code['code'])) {
-                    $survey_link =  $survey_link . '?' . $survey_info['participant_variable'] . '=' . $user_code['code'];
+                $user_code = $this->db->get_user_code();
+                if ($user_code) {
+                    $survey_link =  $survey_link . '?' . $survey_info['participant_variable'] . '=' . $user_code;
                 }
             }
         }